+++ packages/storage-plugin/lib/index.ts
@@ -1,0 +1,2 @@
+export { StoragePlugin } from './plugin';
+export type * from './types';